move help.go from commands subpackage to main cli package
The commands/ subpackage only contained help.go. Moving it back to the
main cli package for consistency and removing the empty subpackage.

Entire-Checkpoint: 8e0a64736bc7
