+++ cmd/entire/cli/help.go
@@ -1,4 +1,4 @@
-package commands
+package cli
 
 import (
 	"fmt"
